split whacklib.c into two files so that unnecessary code is not linked in. Also easier to maintain
