+++ include/whack.h
@@ -114,6 +114,80 @@ enum whack_CBOR_actions {
     WHACK_ADD_KEY  =10,
 };
 
+#define CborSignatureTag 55799
+#define CborOpenSwanTag  0x4f50534e
+#define CborIPv4Tag      260            /* squatted */
+#define CborIPv6Tag      261
+
+/* values < 24 get encoded in one byte, < 256 in two bytes */
+enum whack_cbor_attributes {
+      WHACK_OPT_NAME = 1,
+      WHACK_OPT_DEBUGGING = 2,
+      WHACK_OPT_ASYNC = 128,
+      WHACK_OPT_SET   = 129,
+      WHACK_OPT_RECORDFILE=130,
+      WHACK_OPT_MYID  = 131,
+      WHACK_OPT_DELETE= 27,
+      WHACK_OPT_CRASHPEER=132,
+      WHACK_OPT_LISTEN   =133,
+      WHACK_OPT_UNLISTEN =134,
+      WHACK_OPT_REREAD   =135,
+      WHACK_OPT_LIST     =136,
+      WHACK_OPT_PURGE_OCSP=137,
+      WHACK_OPT_IKE      = 139,
+      WHACK_OPT_ESP      = 140,
+      WHACK_OPT_CONNALIAS= 141,
+      WHACK_OPT_POLICYLABEL=142,
+      WHACK_OPT_OPPO_MY_CLIENT = 143,
+      WHACK_OPT_OPPO_PEER_CLIENT=144,
+      WHACK_OPT_DELETESTATE=145,
+
+      WHACK_OPT_LEFT     = 3,
+      WHACK_OPT_RIGHT    = 4,
+
+      WHACK_OPT_LIFETIME_IKE = 146,
+      WHACK_OPT_LIFETIME_IPSEC=147,
+      WHACK_OPT_LIFETIME_REKEY_MARGIN=148,
+      WHACK_OPT_LIFETIME_REKEY_FUZZ=149,
+      WHACK_OPT_LIFETIME_REKEY_TRIES=150,
+      WHACK_OPT_POLICY        = 151,
+      WHACK_OPT_KEYVAL        = 15,
+      WHACK_OPT_KEYID         = 16,
+      WHACK_OPT_KEYALG        = 17,
+      WHACK_OPT_END_ADDR_FAMILY=18,
+};
+
+enum whack_cbor_end_attr {
+      WHACK_OPT_END_ID   = 5,
+      WHACK_OPT_END_CERT = 6,
+      WHACK_OPT_END_CA   = 7,
+      WHACK_OPT_END_GROUPS =8,
+      WHACK_OPT_END_VIRT = 9,
+      WHACK_OPT_END_XAUTH_NAME =137,       /* uncommon */
+      WHACK_OPT_END_HOST_ADDRNAME = 10,
+      WHACK_OPT_END_HOST_ADDR     = 11,
+      WHACK_OPT_END_HOST_NEXTHOP  = 12,
+      WHACK_OPT_END_HOST_SRCIP    = 13,
+      WHACK_OPT_END_CLIENT        = 14,
+
+      WHACK_OPT_HOST_TYPE = 15,
+      WHACK_OPT_KEYTYPE   = 16,
+      WHACK_OPT_HAS_CLIENT= 17,
+      WHACK_OPT_HAS_CLIENT_WILDCARD=18,
+      WHACK_OPT_HAS_PORT_WILDCARD=19,
+      WHACK_OPT_HOST_PORT=20,
+      WHACK_OPT_PORT=138,
+      WHACK_OPT_XAUTH_SERVER=139,
+      WHACK_OPT_XAUTH_CLIENT=140,
+      WHACK_OPT_MODECFG_SERVER=141,
+      WHACK_OPT_MODECFG_CLIENT=142,
+      WHACK_OPT_CERTPOLICY=143,
+      WHACK_OPT_CERTTYPE=144,
+      WHACK_OPT_TUNDEV=145
+};
+
+
+
 /* this is the historic message from Openswan < 3.1 */
 struct legacy_whack_message {
     u_int32_t magic;
