+++ lib/libwhack/Makefile
@@ -22,7 +22,7 @@ include ${OPENSWANSRCDIR}/Makefile.ver
 MANDIR=$(MANTREE)/man3
 
 QCB=${OPENSWANSRCDIR}/lib/libwhack/qcbor
-SRCS=whacklib.c aliascomp.c
+SRCS=whackread.c whackwrite.c aliascomp.c
 
 SRCS+=UsefulBuf.c qcbor_encode.c qcbor_decode.c qcbor_err_to_str.c
 SRCS+=ieee754.c
@@ -72,8 +72,8 @@ MANS=
 .PHONY:	all install clean l t lt tar check depend checkprograms
 
 # we can't do tests with two conditions, but we can concatenate the strings
-SHOULDWERESTART=${USE_OBJDIR}$(wildcard whacklib.c)
-ifeq ($(SHOULDWERESTART),truewhacklib.c)
+SHOULDWERESTART=${USE_OBJDIR}$(wildcard whackread.c)
+ifeq ($(SHOULDWERESTART),truewhackread.c)
 all programs clean install:
 	cd ${OPENSWANSRCDIR} && cd ${OBJDIRTOP}/lib/libwhack && make $@
 else
