+++ lib/libwhack/whackread.c
@@ -77,381 +77,6 @@ struct whackpacker {
 #endif
 
 #define OK(x) ugh = (x); if(ugh) goto bad
-#define CborSignatureTag 55799
-#define CborOpenSwanTag  0x4f50534e
-#define CborIPv4Tag      260
-#define CborIPv6Tag      261
-
-/* values < 24 get encoded in one byte, < 256 in two bytes */
-enum whack_cbor_attributes {
-      WHACK_OPT_NAME = 1,
-      WHACK_OPT_DEBUGGING = 2,
-      WHACK_OPT_ASYNC = 128,
-      WHACK_OPT_SET   = 129,
-      WHACK_OPT_RECORDFILE=130,
-      WHACK_OPT_MYID  = 131,
-      WHACK_OPT_DELETE= 27,
-      WHACK_OPT_CRASHPEER=132,
-      WHACK_OPT_LISTEN   =133,
-      WHACK_OPT_UNLISTEN =134,
-      WHACK_OPT_REREAD   =135,
-      WHACK_OPT_LIST     =136,
-      WHACK_OPT_PURGE_OCSP=137,
-      WHACK_OPT_IKE      = 139,
-      WHACK_OPT_ESP      = 140,
-      WHACK_OPT_CONNALIAS= 141,
-      WHACK_OPT_POLICYLABEL=142,
-      WHACK_OPT_OPPO_MY_CLIENT = 143,
-      WHACK_OPT_OPPO_PEER_CLIENT=144,
-      WHACK_OPT_DELETESTATE=145,
-
-      WHACK_OPT_LEFT     = 3,
-      WHACK_OPT_RIGHT    = 4,
-
-      WHACK_OPT_LIFETIME_IKE = 146,
-      WHACK_OPT_LIFETIME_IPSEC=147,
-      WHACK_OPT_LIFETIME_REKEY_MARGIN=148,
-      WHACK_OPT_LIFETIME_REKEY_FUZZ=149,
-      WHACK_OPT_LIFETIME_REKEY_TRIES=150,
-      WHACK_OPT_POLICY        = 151,
-      WHACK_OPT_KEYVAL        = 15,
-      WHACK_OPT_KEYID         = 16,
-      WHACK_OPT_KEYALG        = 17,
-      WHACK_OPT_END_ADDR_FAMILY=18,
-};
-
-enum whack_cbor_end_attr {
-      WHACK_OPT_END_ID   = 5,
-      WHACK_OPT_END_CERT = 6,
-      WHACK_OPT_END_CA   = 7,
-      WHACK_OPT_END_GROUPS =8,
-      WHACK_OPT_END_VIRT = 9,
-      WHACK_OPT_END_XAUTH_NAME =137,       /* uncommon */
-      WHACK_OPT_END_HOST_ADDRNAME = 10,
-      WHACK_OPT_END_HOST_ADDR     = 11,
-      WHACK_OPT_END_HOST_NEXTHOP  = 12,
-      WHACK_OPT_END_HOST_SRCIP    = 13,
-      WHACK_OPT_END_CLIENT        = 14,
-
-      WHACK_OPT_HOST_TYPE = 15,
-      WHACK_OPT_KEYTYPE   = 16,
-      WHACK_OPT_HAS_CLIENT= 17,
-      WHACK_OPT_HAS_CLIENT_WILDCARD=18,
-      WHACK_OPT_HAS_PORT_WILDCARD=19,
-      WHACK_OPT_HOST_PORT=20,
-      WHACK_OPT_PORT=138,
-      WHACK_OPT_XAUTH_SERVER=139,
-      WHACK_OPT_XAUTH_CLIENT=140,
-      WHACK_OPT_MODECFG_SERVER=141,
-      WHACK_OPT_MODECFG_CLIENT=142,
-      WHACK_OPT_CERTPOLICY=143,
-      WHACK_OPT_CERTTYPE=144,
-      WHACK_OPT_TUNDEV=145
-};
-
-#if 0
-static void whack_cbor_encode_empty_map(QCBOREncodeContext *qec)
-{
-  QCBOREncode_OpenMap(qec);
-  QCBOREncode_CloseMap(qec);
-}
-#endif
-
-static err_t whack_cbor_magic_header(QCBOREncodeContext *qec)
-{
-  UsefulBufC bor = UsefulBuf_FROM_SZ_LITERAL("BOR");
-  QCBOREncode_AddTag(qec, CborSignatureTag);
-  QCBOREncode_AddTag(qec, CborOpenSwanTag);
-  QCBOREncode_AddBytes(qec, bor);
-  return NULL;
-}
-
-static void whack_cbor_encode_ipaddress(QCBOREncodeContext *qec, ip_address *addr)
-{
-  UsefulBufC ub;
-
-  switch(ip_address_family(addr)) {
-  case AF_INET:
-    QCBOREncode_AddTag(qec, CborIPv4Tag);
-    ub.ptr = (const void *)&addr->u.v4.sin_addr.s_addr;
-    ub.len = 4;
-    QCBOREncode_AddBytes(qec, ub);
-    break;
-  case AF_INET6:
-    QCBOREncode_AddTag(qec, CborIPv6Tag);
-    ub.ptr = (const void *)addr->u.v6.sin6_addr.s6_addr;
-    ub.len = 16;
-    QCBOREncode_AddBytes(qec, ub);
-    break;
-  }
-}
-
-static void whack_cbor_encode_some_ipaddress_ToMapN(QCBOREncodeContext *qec
-                                                   , u_int32_t   link
-                                                   , ip_address *addr)
-{
-  if(!ip_address_isany(addr)) {
-    QCBOREncode_AddInt64(qec, link);
-    whack_cbor_encode_ipaddress(qec, addr);
-  }
-}
-
-static void whack_cbor_encode_some_ipsubnet_ToMapN(QCBOREncodeContext *qec
-                                                  , u_int32_t   link
-                                                  , ip_subnet  *net)
-{
-  ip_address *addr = &net->addr;
-
-  if(!ip_address_isany(addr)) {
-    QCBOREncode_OpenArrayInMapN(qec, link);
-    QCBOREncode_AddInt64(qec, net->maskbits);
-    whack_cbor_encode_ipaddress(qec, addr);
-    QCBOREncode_CloseArray(qec);
-  }
-}
-
-#define ADDIntIfNotZero(qec, tag, value) if(value != 0) QCBOREncode_AddInt64ToMapN(qec,tag,value)
-
-static void whack_cbor_encode_end(QCBOREncodeContext *qec, struct whack_end *we)
-{
-  if(we->id) {
-    QCBOREncode_AddSZStringToMapN(qec, WHACK_OPT_END_ID, we->id);
-  }
-  if(we->cert) {
-    QCBOREncode_AddSZStringToMapN(qec, WHACK_OPT_END_CERT, we->cert);
-  }
-  if(we->ca) {
-    QCBOREncode_AddSZStringToMapN(qec, WHACK_OPT_END_CA, we->ca);
-  }
-  if(we->groups) {
-    QCBOREncode_AddSZStringToMapN(qec, WHACK_OPT_END_GROUPS, we->groups);
-  }
-  if(we->virt) {
-    QCBOREncode_AddSZStringToMapN(qec, WHACK_OPT_END_VIRT, we->virt);
-  }
-  if(we->xauth_name) {
-    QCBOREncode_AddSZStringToMapN(qec, WHACK_OPT_END_XAUTH_NAME, we->xauth_name);
-  }
-  if(we->host_addr_name) {
-    QCBOREncode_AddSZStringToMapN(qec, WHACK_OPT_END_HOST_ADDRNAME, we->host_addr_name);
-  }
-
-  ADDIntIfNotZero(qec, WHACK_OPT_HOST_TYPE, we->host_type);
-  /* host_addr */
-  whack_cbor_encode_some_ipaddress_ToMapN(qec, WHACK_OPT_END_HOST_ADDR
-                                            , &we->host_addr);
-
-  ADDIntIfNotZero(qec, WHACK_OPT_KEYTYPE,   we->keytype);
-  ADDIntIfNotZero(qec, WHACK_OPT_HAS_CLIENT, we->has_client);
-  ADDIntIfNotZero(qec, WHACK_OPT_HAS_CLIENT_WILDCARD, we->has_client_wildcard);
-  ADDIntIfNotZero(qec, WHACK_OPT_HAS_PORT_WILDCARD, we->has_port_wildcard);
-  ADDIntIfNotZero(qec, WHACK_OPT_HOST_PORT, we->host_port);
-  ADDIntIfNotZero(qec, WHACK_OPT_PORT,      we->protocol);
-  ADDIntIfNotZero(qec, WHACK_OPT_XAUTH_SERVER, we->xauth_server);
-  ADDIntIfNotZero(qec, WHACK_OPT_XAUTH_CLIENT, we->xauth_client);
-  ADDIntIfNotZero(qec, WHACK_OPT_MODECFG_SERVER, we->modecfg_server);
-  ADDIntIfNotZero(qec, WHACK_OPT_MODECFG_CLIENT, we->modecfg_client);
-  ADDIntIfNotZero(qec, WHACK_OPT_CERTPOLICY, we->sendcert);
-  ADDIntIfNotZero(qec, WHACK_OPT_CERTTYPE,   we->certtype);
-  ADDIntIfNotZero(qec, WHACK_OPT_TUNDEV,     we->tundev);
-
-  /* host_nexthop */
-  whack_cbor_encode_some_ipaddress_ToMapN(qec, WHACK_OPT_END_HOST_NEXTHOP
-                                         , &we->host_nexthop);
-
-  /* host_srcip */
-  whack_cbor_encode_some_ipaddress_ToMapN(qec, WHACK_OPT_END_HOST_SRCIP
-                                         , &we->host_srcip);
-
-  /* client */
-  whack_cbor_encode_some_ipsubnet_ToMapN(qec, WHACK_OPT_END_CLIENT
-                                       , &we->client);
-
-}
-
-err_t whack_cbor_encode_msg(struct whack_message *wm, unsigned char *buf, size_t *plen)
-{
-  size_t outlen;
-  QCBOREncodeContext qec;
-  err_t ugh= NULL;
-  QCBORError e;
-
-  UsefulBuf into = {buf, (unsigned long)*plen};
-  QCBOREncode_Init(&qec, into);
-
-  OK(whack_cbor_magic_header(&qec));
-
-  QCBOREncode_OpenMap(&qec);
-  if(wm->whack_status) {
-    QCBOREncode_OpenMapInMapN(&qec, WHACK_STATUS);
-    QCBOREncode_CloseMap(&qec);
-  }
-
-  if(wm->whack_shutdown) {
-    QCBOREncode_OpenMapInMapN(&qec, WHACK_SHUTDOWN);
-    QCBOREncode_CloseMap(&qec);
-  }
-
-  if(wm->whack_options) {
-    QCBOREncode_OpenMapInMapN(&qec, WHACK_OPTIONS);
-    QCBOREncode_AddInt64ToMapN(&qec, WHACK_OPT_SET, wm->opt_set);
-    if(wm->name) {
-      QCBOREncode_AddSZStringToMapN(&qec, WHACK_OPT_NAME, wm->name);
-    }
-    if(wm->string1) {
-      QCBOREncode_AddSZStringToMapN(&qec, WHACK_OPT_RECORDFILE, wm->string1);
-    }
-    QCBOREncode_AddInt64ToMapN(&qec, WHACK_OPT_DEBUGGING, wm->debugging);
-    QCBOREncode_CloseMap(&qec);
-  }
-
-  if (wm->whack_connection) {
-    QCBOREncode_OpenMapInMapN(&qec, WHACK_CONNECTION);
-
-    QCBOREncode_OpenMapInMapN(&qec, WHACK_OPT_LEFT);
-    whack_cbor_encode_end(&qec, &wm->left);
-    QCBOREncode_CloseMap(&qec);
-
-    QCBOREncode_OpenMapInMapN(&qec, WHACK_OPT_RIGHT);
-    whack_cbor_encode_end(&qec, &wm->right);
-    QCBOREncode_CloseMap(&qec);
-
-    QCBOREncode_AddInt64ToMapN(&qec, WHACK_OPT_POLICY, wm->policy);
-    QCBOREncode_AddInt64ToMapN(&qec, WHACK_OPT_LIFETIME_IKE, wm->sa_ike_life_seconds);
-    QCBOREncode_AddInt64ToMapN(&qec, WHACK_OPT_LIFETIME_IPSEC, wm->sa_ipsec_life_seconds);
-    QCBOREncode_AddInt64ToMapN(&qec, WHACK_OPT_LIFETIME_REKEY_MARGIN, wm->sa_rekey_margin);
-    QCBOREncode_AddInt64ToMapN(&qec, WHACK_OPT_LIFETIME_REKEY_FUZZ, wm->sa_rekey_fuzz);
-    QCBOREncode_AddInt64ToMapN(&qec, WHACK_OPT_LIFETIME_REKEY_TRIES, wm->sa_keying_tries);
-    QCBOREncode_AddInt64ToMapN(&qec, WHACK_OPT_END_ADDR_FAMILY, wm->end_addr_family);
-
-    QCBOREncode_CloseMap(&qec);
-  }
-
-  if(wm->whack_async) {
-    QCBOREncode_AddInt64ToMapN(&qec, WHACK_OPT_ASYNC, wm->whack_async);
-  }
-
-  if(wm->whack_myid) {
-    QCBOREncode_AddSZStringToMapN(&qec, WHACK_OPT_MYID, wm->myid);
-  }
-
-  if(wm->whack_delete) {
-    QCBOREncode_AddSZStringToMapN(&qec, WHACK_OPT_DELETE, wm->name);
-  }
-
-  if(wm->whack_deletestate) {
-    QCBOREncode_AddInt64ToMapN(&qec, WHACK_OPT_DELETESTATE, wm->whack_deletestateno);
-  }
-
-  if(wm->whack_crash) {
-    /* open code the IPAddressToMap */
-    QCBOREncode_AddInt64(&qec, WHACK_OPT_CRASHPEER);
-    whack_cbor_encode_ipaddress(&qec, &wm->whack_crash_peer);
-  }
-
-  if(wm->whack_listen) {
-    QCBOREncode_AddInt64ToMapN(&qec, WHACK_OPT_LISTEN, 1);
-  }
-  if(wm->whack_unlisten) {
-    QCBOREncode_AddInt64ToMapN(&qec, WHACK_OPT_UNLISTEN, 1);
-  }
-  if(wm->whack_reread) {
-    QCBOREncode_AddInt64ToMapN(&qec, WHACK_OPT_REREAD, wm->whack_reread);
-  }
-  if(wm->whack_list) {
-    QCBOREncode_AddInt64ToMapN(&qec, WHACK_OPT_LIST, wm->whack_list);
-  }
-  if(wm->whack_purgeocsp) {
-    QCBOREncode_AddInt64ToMapN(&qec, WHACK_OPT_PURGE_OCSP, wm->whack_purgeocsp);
-  }
-
-  if(wm->whack_route) {
-    QCBOREncode_OpenMapInMapN(&qec, WHACK_ROUTE);
-    if(wm->name) {
-      QCBOREncode_AddSZStringToMapN(&qec, WHACK_OPT_NAME, wm->name);
-    }
-    QCBOREncode_CloseMap(&qec);
-  }
-
-  if(wm->whack_unroute) {
-    QCBOREncode_OpenMapInMapN(&qec, WHACK_UNROUTE);
-    if(wm->name) {
-      QCBOREncode_AddSZStringToMapN(&qec, WHACK_OPT_NAME, wm->name);
-    }
-    QCBOREncode_CloseMap(&qec);
-  }
-  if(wm->whack_initiate) {
-    QCBOREncode_OpenMapInMapN(&qec, WHACK_INITIATE);
-    if(wm->name) {
-      QCBOREncode_AddSZStringToMapN(&qec, WHACK_OPT_NAME, wm->name);
-    }
-    QCBOREncode_CloseMap(&qec);
-  }
-  if(wm->whack_oppo_initiate) {
-    QCBOREncode_OpenMapInMapN(&qec, WHACK_INITIATE_OPPO);
-    whack_cbor_encode_some_ipaddress_ToMapN(&qec, WHACK_OPT_OPPO_MY_CLIENT, &wm->oppo_my_client);
-    whack_cbor_encode_some_ipaddress_ToMapN(&qec, WHACK_OPT_OPPO_PEER_CLIENT, &wm->oppo_peer_client);
-    QCBOREncode_CloseMap(&qec);
-  }
-
-  if(wm->whack_terminate) {
-    QCBOREncode_OpenMapInMapN(&qec, WHACK_TERMINATE);
-    if(wm->name) {
-      QCBOREncode_AddSZStringToMapN(&qec, WHACK_OPT_NAME, wm->name);
-    }
-    QCBOREncode_CloseMap(&qec);
-  }
-
-  if(wm->ike) {
-    QCBOREncode_AddSZStringToMapN(&qec, WHACK_OPT_IKE, wm->ike);
-  }
-  if(wm->esp) {
-    QCBOREncode_AddSZStringToMapN(&qec, WHACK_OPT_ESP, wm->esp);
-  }
-  if(wm->connalias) {
-    QCBOREncode_AddSZStringToMapN(&qec, WHACK_OPT_CONNALIAS, wm->connalias);
-  }
-
-  if(wm->policy_label) {
-    QCBOREncode_AddSZStringToMapN(&qec, WHACK_OPT_POLICYLABEL, wm->policy_label);
-  }
-
-  if(wm->whack_key) {
-    QCBOREncode_OpenMapInMapN(&qec, WHACK_ADD_KEY);
-
-    if(wm->keyid) {
-      QCBOREncode_AddSZStringToMapN(&qec, WHACK_OPT_KEYID, wm->keyid);
-    }
-
-    if(wm->keyval.ptr && wm->keyval.len > 0) {
-      UsefulBufC ub;
-      ub.ptr = wm->keyval.ptr;
-      ub.len = wm->keyval.len;
-      QCBOREncode_AddBytesToMapN(&qec, WHACK_OPT_KEYVAL, ub);
-    }
-    QCBOREncode_AddInt64ToMapN(&qec, WHACK_OPT_KEYALG, wm->pubkey_alg);
-    QCBOREncode_CloseMap(&qec);
-  }
-
-  QCBOREncode_CloseMap(&qec);
-
-  /* close the array */
-  e = QCBOREncode_FinishGetSize(&qec, &outlen);
-  if(e != QCBOR_SUCCESS) {
-    ugh = "encoding failed";
-    return ugh;
-  }
-
-  if(plen) {
-    *plen = outlen;
-  }
-  return NULL;
-
- bad:
-  return "CBOR encoding error";
-}
-
-
 /*
 ** DECODING
 **
@@ -1098,68 +723,5 @@ clear_end(struct whack_end *e)
     e->ca = NULL;
     e->updown = NULL;
     e->host_port = IKE_UDP_PORT;
-}
-
-int
-whack_get_value(char *buf, size_t bufsize)
-{
-    int len;
-    int try;
-
-    fflush(stdout);
-    usleep(20000); /* give fflush time for flushing - has to go through awk */
-
-    try = 3;
-    len = 0;
-    while(try > 0 && len==0)
-    {
-	fprintf(stderr, "Enter username:   ");
-
-	memset(buf, 0, bufsize);
-
-	if(fgets(buf, bufsize, stdin) != buf) {
-	    if(errno == 0) {
-		fprintf(stderr, "Can not read password from standard in\n");
-		exit(RC_WHACK_PROBLEM);
-	    } else {
-		perror("fgets value");
-		exit(RC_WHACK_PROBLEM);
-	    }
-	}
-
-	/* send the value to pluto, including \0, but fgets adds \n */
-	len = strlen(buf);
-	if(len == 0)
-	{
-	    fprintf(stderr, "answer was empty, retry\n");
-	}
-
-        try--;
-    }
-
-    if(len ==  0)
-    {
-	exit(RC_WHACK_PROBLEM);
-    }
-
-    return len;
-}
-
-size_t
-whack_get_secret(char *buf, size_t bufsize)
-{
-    const char *secret;
-    int len;
-
-    fflush(stdout);
-    usleep(20000); /* give fflush time for flushing */
-    secret = getpass("Enter passphrase: ");
-    secret = (secret == NULL) ? "" : secret;
-
-    strncpy(buf, secret, bufsize);
-
-    len = strlen(buf) + 1;
-
-    return len;
 }
